Security: fix use of child_process.exec
exec is run in a shell context and thus parses bash syntax.
This leads for the possibility for running unintended code by crafting malicious links.
execFile is preferable in this scenario by passing args directly and avoiding bash.

However, there is no reason to believe that this security flaw was exploitable.
User crafted links were tested to not be suceptible to this flaw.
Links which where susceptible to this flaw (e.g, Dropbox authentication)
were generated by the app itself which is not malicious or otherwise flawed.